fix(frontend): テーマ配線を管理コンソールの内側へ閉じ込める
店舗ドメインは公開サイトと管理コンソールを同一 origin で配信するため、
provider を根 layout に置くと <html> の class と color-scheme が公開
サイトにも及ぶ。実測では body の背景が白から近黒へ反転していた。

provider は自身が搭載されている間だけ <html> を書き換えるので、
コンソール layout へ下ろせば公開サイト側は素のまま残る。錨は「根には
無い」と「コンソールには規定の props で有る」を対で固定する。
